requirements.txt
et_xmlfile==2.0.0
greenlet==3.2.1
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
xlrd==1.2.0
